# docs/openapi-semana9.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MiniMarket Plus API",
    "description": "API REST de MiniMarket Plus para autenticación, catálogo, carritos, inventario, ventas, sucursales, stock por sucursal, proveedores, órdenes de compra, pedidos, promociones y reportes. Protegida mediante JWT y enriquecida con enlaces HATEOAS.",
    "contact": {
      "name": "Equipo MiniMarket Plus"
    },
    "version": "2.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Servidor local"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/sucursales/{id}": {
      "get": {
        "tags": [
          "Sucursales"
        ],
        "operationId": "obtener",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "description": "Identificador de la sucursal",
            "example": 1
          }
        ],
        "responses": {
          "200": {
            "description": "Sucursal encontrada",
            "content": {
              "application/hal+json": {
                "schema": {
                  "$ref": "#/components/schemas/EntityModelSucursal"
                }
              }
            }
          },
          "401": {
            "description": "Token ausente o inválido"
          },
          "403": {
            "description": "Rol no autorizado"
          },
          "404": {
            "description": "Sucursal no encontrada"
          }
        },
        "summary": "Obtener sucursal",
        "description": "Obtiene una sucursal por su identificador."
      },
      "put": {
        "tags": [
          "Sucursales"
        ],
        "operationId": "actualizar",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "description": "Identificador de la sucursal",
            "example": 1
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Sucursal"
              }
            }
          },
          "required": true,
          "description": "Datos actualizados de la sucursal"
        },
        "responses": {
          "200": {
            "description": "Sucursal actualizada",
            "content": {
              "application/hal+json": {
                "schema": {
                  "$ref": "#/components/schemas/EntityModelSucursal"
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos"
          },
          "401": {
            "description": "Token ausente o inválido"
          },
          "403": {
            "description": "Se requiere ROLE_ADMIN"
          },
          "404": {
            "description": "Sucursal no encontrada"
          },
          "409": {
            "description": "Nombre de sucursal duplicado"
          }
        },
        "summary": "Actualizar sucursal",
        "description": "Actualiza los datos de una sucursal existente. Requiere ROLE_ADMIN."
      },
      "delete": {
        "tags": [
          "Sucursales"
        ],
        "operationId": "eliminar",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "description": "Identificador de la sucursal",
            "example": 1
          }
        ],
        "responses": {
          "204": {
            "description": "Sucursal eliminada"
          },
          "401": {
            "description": "Token ausente o inválido"
          },
          "403": {
            "description": "Se requiere ROLE_ADMIN"
          },
          "404": {
            "description": "Sucursal no encontrada"
          },
          "409": {
            "description": "La sucursal posee relaciones activas"
          }
        },
        "summary": "Eliminar sucursal",
        "description": "Elimina una sucursal existente. Requiere ROLE_ADMIN."
      }
    },
    "/api/proveedores/{id}": {
      "get": {
        "tags": [
          "Proveedores"
        ],
        "operationId": "obtener_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "description": "Identificador del proveedor",
            "example": 1
          }
        ],
        "responses": {
          "200": {
            "description": "Proveedor encontrado",
            "content": {
              "application/hal+json": {
                "schema": {
                  "$ref": "#/components/schemas/EntityModelProveedor"
                }
              }
            }
          },
          "401": {
            "description": "Token ausente o inválido"
          },
          "403": {
            "description": "Se requiere ROLE_ADMIN"
          },
          "404": {
            "description": "Proveedor no encontrado"
          }
        },
        "summary": "Obtener proveedor",
        "description": "Obtiene un proveedor por su identificador. Requiere ROLE_ADMIN."
      },
      "put": {
        "tags": [
          "Proveedores"
        ],
        "operationId": "actualizar_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "description": "Identificador del proveedor",
            "example": 1
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Proveedor"
              }
            }
          },
          "required": true,
          "description": "Datos actualizados del proveedor"
        },
        "responses": {
          "200": {
            "description": "Proveedor actualizado",
            "content": {
              "application/hal+json": {
                "schema": {
                  "$ref": "#/components/schemas/EntityModelProveedor"
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos"
          },
          "401": {
            "description": "Token ausente o inválido"
          },
          "403": {
            "description": "Se requiere ROLE_ADMIN"
          },
          "404": {
            "description": "Proveedor no encontrado"
          },
          "409": {
            "description": "Correo de proveedor duplicado"
          }
        },
        "summary": "Actualizar proveedor",
        "description": "Actualiza un proveedor existente. Requiere ROLE_ADMIN."
      },
      "delete": {
        "tags": [
          "Proveedores"
        ],
        "operationId": "eliminar_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "description": "Identificador del proveedor",
            "example": 1
          }
        ],
        "responses": {
          "204": {
            "description": "Proveedor eliminado"
          },
          "401": {
            "description": "Token ausente o inválido"
          },
          "403": {
            "description": "Se requiere ROLE_ADMIN"
          },
          "404": {
            "description": "Proveedor no encontrado"
          },
          "409": {
            "description": "El proveedor posee órdenes relacionadas"
          }
        },
        "summary": "Eliminar proveedor",
        "description": "Elimina un proveedor existente. Requiere ROLE_ADMIN."
      }
    },
    "/api/promociones/{id}": {
      "get": {
        "tags": [
          "Promociones"
        ],
        "operationId": "obtener_2",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "description": "Identificador de la promoción",
            "example": 1
          }
        ],
        "responses": {
          "200": {
            "description": "Promoción encontrada",
            "content": {
              "application/hal+json": {
                "schema": {
                  "$ref": "#/components/schemas/EntityModelPromocion"
                }
              }
            }
          },
          "401": {
            "description": "Token ausente o inválido"
          },
          "403": {
            "description": "Se requiere ROLE_ADMIN"
          },
          "404": {
            "description": "Promoción no encontrada"
          }
        },
        "summary": "Obtener promoción",
        "description": "Obtiene una promoción por su identificador. Requiere ROLE_ADMIN."
      },
      "put": {
        "tags": [
          "Promociones"
        ],
        "operationId": "actualizar_2",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "description": "Identificador de la promoción",
            "example": 1
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PromocionRequest"
              }
            }
          },
          "required": true,
          "description": "Datos actualizados de la promoción"
        },
        "responses": {
          "200": {
            "description": "Promoción actualizada",
            "content": {
              "application/hal+json": {
                "schema": {
                  "$ref": "#/components/schemas/EntityModelPromocion"
                }
              }
            }
          },
          "400": {
            "description": "Datos o fechas inválidos"
          },
          "401": {
            "description": "Token ausente o inválido"
          },
          "403": {
            "description": "Se requiere ROLE_ADMIN"
          },
          "404": {
            "description": "Promoción o producto no encontrado"
          }
        },
        "summary": "Actualizar promoción",
        "description": "Actualiza una promoción existente. Requiere ROLE_ADMIN."
      },
      "delete": {
        "tags": [
          "Promociones"
        ],
        "operationId": "eliminar_2",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "description": "Identificador de la promoción",
            "example": 1
          }
        ],
        "responses": {
          "204": {
            "description": "Promoción eliminada"
          },
          "401": {
            "description": "Token ausente o inválido"
          },
          "403": {
            "description": "Se requiere ROLE_ADMIN"
          },
          "404": {
            "description": "Promoción no encontrada"
          }
        },
        "summary": "Eliminar promoción",
        "description": "Elimina una promoción existente. Requiere ROLE_ADMIN."
      }
    },
    "/api/productos/{id}": {
      "get": {
        "tags": [
          "Productos"
        ],
        "summary": "Obtener producto",
        "description": "Busca un producto por su identificador.",
        "operationId": "obtenerProductoPorId",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Identificador del producto",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 1
          }
        ],
        "responses": {
          "200": {
            "description": "Producto encontrado",
            "content": {
              "application/hal+json": {
                "schema": {
                  "$ref": "#/components/schemas/Producto"
                },
                "example": {
                  "id": 1,
                  "nombre": "Arroz 1 kg",
                  "precio": 1890.0,
                  "stock": 40,
                  "categoria": {
                    "id": 1,
                    "nombre": "Abarrotes"
                  },
                  "_links": {
                    "self": {
                      "href": "http://localhost:8080/api/productos/1"
                    },
                    "productos": {
                      "href": "http://localhost:8080/api/productos"
                    },
                    "categoria": {
                      "href": "http://localhost:8080/api/categorias/1"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Producto no encontrado"
          }
        }
      },
      "put": {
        "tags": [
          "Productos"
        ],
        "summary": "Actualizar producto",
        "description": "Reemplaza los datos de un producto existente.",
        "operationId": "actualizarProducto",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Identificador del producto",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 1
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Producto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Producto actualizado",
            "content": {
              "application/hal+json": {
                "schema": {
                  "$ref": "#/components/schemas/EntityModelProducto"
                }
              }
            }
          },
          "404": {
            "description": "Producto no encontrado"
          },
          "403": {
            "description": "El usuario no posee rol administrador"
          }
        }
      },
      "delete": {
        "tags": [
          "Productos"
        ],
        "summary": "Eliminar producto",
        "description": "Elimina un producto del catálogo.",
        "operationId": "eliminarProducto",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Identificador del producto",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 1
          }
        ],
        "responses": {
          "204": {
            "description": "Producto eliminado"
          },
          "404": {
            "description": "Producto no encontrado"
          },
          "403": {
            "description": "El usuario no posee rol administrador"
          }
        }
      }
    },
    "/api/pedidos/{id}/estado": {
      "put": {
        "tags": [
          "Pedidos en línea"
        ],
        "operationId": "estado",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "description": "Identificador del pedido",
            "example": 1
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PedidoEstadoRequest"
              }
            }
          },
          "required": true,
          "description": "Nuevo estado del pedido"
        },
        "responses": {
          "200": {
            "description": "Estado actualizado",
            "content": {
              "application/hal+json": {
                "schema": {
                  "$ref": "#/components/schemas/EntityModelPedido"
                }
              }
            }
          },
          "400": {
            "description": "Estado inválido"
          },
          "401": {
            "description": "Token ausente o inválido"
          },
          "403": {
            "description": "Se requiere ROLE_ADMIN"
          },
          "404": {
            "description": "Pedido no encontrado"
          }
        },
        "summary": "Actualizar estado de pedido",
        "description": "Actualiza el estado operativo de un pedido. Requiere ROLE_ADMIN."
      }
    },
...
    "/api/stock-sucursales": {
      "get": {
        "tags": [
          "Stock por sucursal"
        ],
        "operationId": "listar_1",
        "responses": {
          "200": {
            "description": "Stock obtenido correctamente",
            "content": {
              "application/hal+json": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionModelEntityModelStockSucursal"
                }
              }
            }
          },
          "401": {
            "description": "Token ausente o inválido"
          },
          "403": {
            "description": "Rol no autorizado"
          }
        },
        "summary": "Listar stock por sucursal",
        "description": "Obtiene todas las existencias configuradas por sucursal y producto."
      },
      "post": {
        "tags": [
          "Stock por sucursal"
        ],
        "operationId": "registrar",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StockSucursalRequest"
              }
            }
          },
          "required": true,
          "description": "Sucursal, producto, cantidad, tipo de movimiento y stock mínimo opcional"
        },
        "responses": {
          "201": {
            "description": "Movimiento registrado",
            "content": {
              "application/hal+json": {
                "schema": {
                  "$ref": "#/components/schemas/EntityModelStockSucursal"
                }
              }
            }
          },
          "400": {
            "description": "Datos, cantidad o tipo de movimiento inválidos"
          },
          "401": {
            "description": "Token ausente o inválido"
          },
          "403": {
            "description": "Se requiere ROLE_ADMIN"
          },
          "404": {
            "description": "Sucursal o producto no encontrado"
          },
          "409": {
            "description": "Stock insuficiente"
          }
        },
        "summary": "Registrar movimiento de stock por sucursal",
        "description": "Registra una ENTRADA o SALIDA. Una salida que alcanza el mínimo puede generar una orden automática. Requiere ROLE_ADMIN."
      }
    },
...
    "/api/pedidos": {
      "get": {
        "tags": [
          "Pedidos en línea"
        ],
        "operationId": "listar_4",
        "responses": {
          "200": {
            "description": "Pedidos obtenidos correctamente",
            "content": {
              "application/hal+json": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionModelEntityModelPedido"
                }
              }
            }
          },
          "401": {
            "description": "Token ausente o inválido"
          },
          "403": {
            "description": "Rol no autorizado"
          }
        },
        "summary": "Listar pedidos",
        "description": "ROLE_CLIENTE obtiene sus pedidos; ROLE_ADMIN obtiene todos."
      },
      "post": {
        "tags": [
          "Pedidos en línea"
        ],
        "operationId": "crear_3",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PedidoRequest"
              }
            }
          },
          "required": true,
          "description": "Sucursal, tipo de entrega, dirección cuando corresponda y detalles del pedido"
        },
        "responses": {
          "201": {
            "description": "Pedido creado",
            "content": {
              "application/hal+json": {
                "schema": {
                  "$ref": "#/components/schemas/EntityModelPedido"
                }
              }
            }
          },
          "400": {
            "description": "Datos, cantidades o dirección inválidos"
          },
          "401": {
            "description": "Token ausente o inválido"
          },
          "403": {
            "description": "Rol no autorizado"
          },
          "404": {
            "description": "Sucursal, producto o stock no encontrado"
          },
          "409": {
            "description": "Stock insuficiente"
          }
        },
        "summary": "Crear pedido",
        "description": "Crea un pedido para retiro o despacho, aplica promociones vigentes y descuenta stock por sucursal."
      }
    },
...
    "/api/pedidos/{id}": {
      "get": {
        "tags": [
          "Pedidos en línea"
        ],
        "operationId": "obtener_3",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "description": "Identificador del pedido",
            "example": 1
          }
        ],
        "responses": {
          "200": {
            "description": "Pedido encontrado",
            "content": {
              "application/hal+json": {
                "schema": {
                  "$ref": "#/components/schemas/EntityModelPedido"
                }
              }
            }
          },
          "401": {
            "description": "Token ausente o inválido"
          },
          "403": {
            "description": "Rol no autorizado"
          },
          "404": {
            "description": "Pedido no encontrado"
          }
        },
        "summary": "Obtener pedido",
        "description": "ROLE_CLIENTE solo puede consultar pedidos propios; ROLE_ADMIN puede consultar cualquiera."
      }
    },
    "/api/pedidos/{id}/detalles": {
      "get": {
        "tags": [
          "Pedidos en línea"
        ],
        "operationId": "detalles",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "description": "Identificador del pedido",
            "example": 1
          }
        ],
        "responses": {
          "200": {
            "description": "Detalles obtenidos correctamente",
            "content": {
              "application/hal+json": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionModelEntityModelDetallePedido"
                }
              }
            }
          },
          "401": {
            "description": "Token ausente o inválido"
          },
          "403": {
            "description": "Rol no autorizado"
          },
          "404": {
            "description": "Pedido no encontrado"
          }
        },
        "summary": "Listar detalles de un pedido",
        "description": "Obtiene los productos y precios asociados a un pedido accesible para el usuario autenticado."
      }
    },
...
    "/api/ordenes-compra/{id}": {
      "get": {
        "tags": [
          "Órdenes de compra"
        ],
        "operationId": "obtener_4",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "description": "Identificador de la orden",
            "example": 1
          }
        ],
        "responses": {
          "200": {
            "description": "Orden encontrada",
            "content": {
              "application/hal+json": {
                "schema": {
                  "$ref": "#/components/schemas/EntityModelOrdenCompra"
                }
              }
            }
          },
          "401": {
            "description": "Token ausente o inválido"
          },
          "403": {
            "description": "Se requiere ROLE_ADMIN"
          },
          "404": {
            "description": "Orden no encontrada"
          }
        },
        "summary": "Obtener orden de compra",
        "description": "Obtiene una orden por su identificador. Requiere ROLE_ADMIN."
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Sucursal": {
        "type": "object",
        "description": "Sucursal de MiniMarket Plus.",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "readOnly": true
          },
          "nombre": {
            "type": "string"
          },
          "direccion": {
            "type": "string"
          },
          "activa": {
            "type": "boolean"
          }
        }
      },
      "EntityModelSucursal": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "readOnly": true
          },
          "nombre": {
            "type": "string"
          },
          "direccion": {
            "type": "string"
          },
          "activa": {
            "type": "boolean"
          },
          "_links": {
            "$ref": "#/components/schemas/Links"
          }
        }
      },
...
      "PedidoEstadoRequest": {
        "type": "object",
        "description": "Cambio de estado administrativo de un pedido.",
        "properties": {
          "estado": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "Ingrese únicamente el token generado por POST /api/auth/login.",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}